backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
alembic==1.13.1
firebase-admin==6.3.0
requests==2.31.0
pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0
beautifulsoup4==4.12.3
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.5
